fix for alpha: size_t and unsigned are not compatible
git-svn-id: svn://svn.code.sf.net/p/synce/code/trunk@130 599ab11a-0111-0410-8a50-d7a5b0953570

+++ librapi2/src/rapi_buffer.c
@@ -68,7 +68,7 @@ bool rapi_buffer_reset(RapiBuffer* buffer, unsigned char* data, size_t size)
 	return true;
 }
 
-unsigned rapi_buffer_get_size(RapiBuffer* buffer)
+size_t rapi_buffer_get_size(RapiBuffer* buffer)
 {
 	return buffer->bytes_used;
 }
